andreo_tsyki — копия/andreo_phone/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-anymail==9.0
django-filter==22.1
django-import-export==3.0.2
django-phone-field==1.8.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-resized==1.0.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.1
phonenumbers==8.13.4
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tablib==3.3.0
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
zipp==3.12.0
